requirements.txt
backtrader>=1.9.78.123
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
PyYAML>=6.0
PyMySQL>=1.1.0
PyJWT>=2.8.0
cryptography>=41.0.0
